Hot fix

+++ HomeCleanAPI/src/main/java/com/example/homecleanapi/repositories/CustomerRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface CustomerRepository extends JpaRepository<Customers, Long> {
     Customers findByPhone(String phone);
-    Customers findById(long id);
+    Customers findById(Integer id);
     boolean existsByPhone(String phone);
 }
